fix: remove metric units that are interpreted as ratios (#49)
Metrics with a unit of "1" are interpreted as ratios by the prometheus
exporter, which has the side effect of appending `_ratio` to the metric
name. This is supposed to only be done for gauges but is actually
applied to counters too. See
https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/pkg/translator/prometheus
for the transformations applied by the prometheus exporter. Since we
were using "1" for dimensionless counts I have simply removed the units.

This PR also hooks up the SentMessages, SentMessageErrors, SentRequests,
SentRequestErrors, OutboundRequestLatency and SentBytes DHT metrics.